This commit was generated by cvs2svn to compensate for changes in r15971, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@15972 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/petcat.c
@@ -976,7 +976,7 @@ static int p_expand(int version, int addr, int ctrls)
 
         /* prevent list protection from terminating listing */
 
-        while (((int)c = getc(source)) != EOF && !c);
+        while ((c = getc(source)) != EOF && !c);
 
         if (c == 0x12 && !line[2] && !line[3]) {  /* 00 00 12 22 */
             directory++;
@@ -1071,7 +1071,7 @@ static int p_expand(int version, int addr, int ctrls)
 
             _p_toascii((int)c, ctrls);  /* convert character */
 
-        } while (((int)c = getc(source)) != EOF && c);
+        } while ((c = getc(source)) != EOF && c);
         fprintf(dest, "\n");
 
     }      /* line */
